+++ tests/testthat/test-validate_pr.R
@@ -69,4 +69,64 @@ test_that("validate_pr flags modifications and deletions in PR", {
   expect_error(
     suppressMessages(check_for_errors(checks))
   )
+
+  checks <- suppressMessages(
+    validate_pr(
+      hub_path = temp_hub,
+      gh_repo = "Infectious-Disease-Modeling-Hubs/ci-testhub-simple",
+      pr_number = 6,
+      skip_submit_window_check = TRUE,
+      file_modify_check = "warn"
+    )
+  )
+  expect_snapshot(str(checks))
+  expect_error(
+    suppressMessages(check_for_errors(checks))
+  )
+
+  checks <- suppressMessages(
+    validate_pr(
+      hub_path = temp_hub,
+      gh_repo = "Infectious-Disease-Modeling-Hubs/ci-testhub-simple",
+      pr_number = 6,
+      skip_submit_window_check = TRUE,
+      file_modify_check = "message"
+    )
+  )
+  expect_snapshot(str(checks))
+  expect_true(
+    suppressMessages(check_for_errors(checks[1:5]))
+  )
+
+
+  checks <- suppressMessages(
+    validate_pr(
+      hub_path = temp_hub,
+      gh_repo = "Infectious-Disease-Modeling-Hubs/ci-testhub-simple",
+      pr_number = 6,
+      skip_submit_window_check = TRUE,
+      file_modify_check = "none"
+    )
+  )
+  expect_snapshot(str(checks))
+  expect_true(
+    suppressMessages(check_for_errors(checks[1:5]))
+  )
+
+  mockery::stub(
+    check_submission_time,
+    "Sys.time",
+    lubridate::as_datetime("2022-10-08 18:01:00 EEST"),
+    2
+  )
+  checks <- suppressMessages(
+    validate_pr(
+      hub_path = temp_hub,
+      gh_repo = "Infectious-Disease-Modeling-Hubs/ci-testhub-simple",
+      pr_number = 6,
+      skip_submit_window_check = TRUE,
+      allow_submit_window_mods = TRUE
+    )
+  )
+  expect_snapshot(str(checks))
 })